% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_immune_space.R
\name{theme_immune_space}
\alias{theme_immune_space}
\title{theme_immune_space}
\usage{
theme_immune_space()
}
\description{
Modifies graphs to recreate the colors and format of
ImmuneSpace graphs
}
\examples{
library(ggplot2)
#Violinplot
p <- ggplot(mtcars, aes(factor(cyl), mpg)) + ggtitle("Violin Plot") + theme_immune_space() + geom_violin()
p + geom_jitter()
}

